pyproject.toml
[tool.poetry]
name = "droid-please"
version = "0.0.1"
description = "Simple cli for genai project management"
authors = ["Luke Lalor <lukehlalor@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
droid = "droid_please.main:app"

[tool.poetry.dependencies]
python = "^3.10"
anthropic = "^0.42.0"
typer = "^0.15.1"
rich = "^13.9.4"
pydantic = "^2.10.4"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
jsonschema = "^4.23.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
ruff = "^0.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
